created ParserTests for all test files
