+++ src/test/java/cmjava2023/helloworld/ASTTest.java
@@ -10,15 +10,17 @@ import org.antlr.v4.runtime.tree.ParseTree;
 import org.cmjava2023.ast.ASTNodes;
 import org.cmjava2023.ast.ASTVisitor;
 import org.cmjava2023.generated_from_antlr.MainAntlrParser;
+import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+@DisplayNameGeneration(cmjava2023.util.QualifiedDisplayNameGenerator.class)
 public class ASTTest {
     @Test
-    public void helloWorld_AST() throws IOException {
+    public void snapshot() throws IOException {
         CharStream charStreamOfGivenFilePath = CharStreams.fromFileName(new TestPathsHelper(this).GetPathOfMainJavaTestResourceInSamePackage());
         Lexer lexer = new org.cmjava2023.generated_from_antlr.MainAntlrLexer(charStreamOfGivenFilePath);
         MainAntlrParser parser = new MainAntlrParser(new CommonTokenStream(lexer));
