+++ src/test/java/cmjava2023/primitive/floats/conversions/ParserTest.java
@@ -1,4 +1,4 @@
-package cmjava2023.helloworld;
+package cmjava2023.primitive.floats.conversions;
 
 import cmjava2023.util.TestPathsHelper;
 import cmjava2023.util.TreePrinter;
