+++ src/test/resources/java-test-files/cmjava2023/primitive/assignment/Main.java
@@ -2,26 +2,44 @@ package cmjava2023.primitive.assignment;
 
 public class Main {
     public static void main(String[] args) {
+        int integer = 8;
+        System.out.println("integer: " + integer);
+        integer = 9;
+        System.out.println("integer: " + integer);
 
+        boolean bool = false;
+        System.out.println("bool: " + bool);
+        bool = true;
+        System.out.println("bool: " + bool);
 
-        int i = 8;
-        boolean value = false;
-        float f = 0.88f;
-        long l = 8888L;
-        char c = '8';
-        byte b = 88;
-        short s = 888;
-        double d = 0.888888888f;
+        float floatingPoint = 0.88f;
+        System.out.println("floatingPoint: " + floatingPoint);
+        floatingPoint = 0.99f;
+        System.out.println("floatingPoint: " + floatingPoint);
 
-        i = 9;
-        value = true;
-        f = 0.99f;
-        l = 99999L;
-        c = '9';
-        b = 99;
-        s = 9999;
-        d = 0.999999999f;
+        long longVariable = 8888L;
+        System.out.println("longVariable: " + longVariable);
+        longVariable = 99999L;
+        System.out.println("longVariable: " + longVariable);
 
-        System.out.println("Hello world!");
+        char character = '8';
+        System.out.println("character: " + character);
+        character = '9';
+        System.out.println("character: " + character);
+
+        byte byteVariable = 88;
+        System.out.println("byteVariable: " + byteVariable);
+        byteVariable = 99;
+        System.out.println("byteVariable: " + byteVariable);
+
+        short shortVariable = 888;
+        System.out.println("shortVariable: " + shortVariable);
+        shortVariable = 9999;
+        System.out.println("shortVariable: " + shortVariable);
+
+        double doubleVariable = 0.888888888f;
+        System.out.println("doubleVariable: " + doubleVariable);
+        doubleVariable = 0.999999999f;
+        System.out.println("doubleVariable: " + doubleVariable);
     }
 }
