--- src/test/resources/java-test-files/cmjava2023/primitive/definition/Main.java
@@ -1,18 +1,0 @@
-package cmjava2023.primitive;
-
-public class Main {
-    public static void main(String[] args) {
-
-
-        int i = 8;
-        boolean value = false;
-        float f = 0.88f;
-        long l = 8888L;
-        char c = '8';
-        byte b = 88;
-        short s = 888;
-        double d = 0.888888888f;
-
-        System.out.println("Hello world!");
-    }
-}
