+++ src/test/resources/java-test-files/cmjava2023/primitive/doubles/mathops/Main.java
@@ -1,0 +1,40 @@
+package cmjava2023.primitive;
+
+public class Main {
+    public static void main(String[] args) {
+        double d = 10;
+        // dadd
+        double d2 = d + d;
+        // ddiv
+        double d3 = d2 / 2;
+        // dmul
+        double d4 = d2 * 2;
+        // dneg
+        double d5 = -d2;
+        // drem
+        double d6 = d2 % 10;
+        // dsub
+        double d7 = d - d;
+
+        System.out.println("d:");
+        System.out.println(d);
+
+        System.out.println("d2:");
+        System.out.println(d2);
+
+        System.out.println("d3:");
+        System.out.println(d3);
+
+        System.out.println("d4:");
+        System.out.println(d4);
+
+        System.out.println("d5:");
+        System.out.println(d5);
+
+        System.out.println("d6:");
+        System.out.println(d6);
+
+        System.out.println("d7:");
+        System.out.println(d7);
+    }
+}
