--- src/test/resources/java-test-files/cmjava2023/primitive/float/mathops/Main.java
@@ -1,40 +1,0 @@
-package org.cmjava2023;
-
-public class Main {
-    public static void main(String[] args) {
-        float f = 10;
-        // ladd
-        float f2 = f + f;
-        // ldiv
-        float f3 = f2 / 2;
-        // lmul
-        float f4 = f2 * 2;
-        // lneg
-        float f5 = -f2;
-        // lrem
-        float f6 = f2 % 10;
-        // lsub
-        float f7 = f - f;
-
-        System.out.println("f:");
-        System.out.println(f);
-
-        System.out.println("f2:");
-        System.out.println(f2);
-
-        System.out.println("f3:");
-        System.out.println(f3);
-
-        System.out.println("f4:");
-        System.out.println(f4);
-
-        System.out.println("f5:");
-        System.out.println(f5);
-
-        System.out.println("f6:");
-        System.out.println(f6);
-
-        System.out.println("f7:");
-        System.out.println(f7);
-    }
-}
