+++ src/test/resources/java-test-files/cmjava2023/primitive/ints/mathops/Main.java
@@ -1,0 +1,46 @@
+package cmjava2023.primitive;
+
+public class Main {
+    public static void main(String[] args) {
+        int i = 10;
+        // iadd
+        int i2 = i + i;
+        // idiv
+        int i3 = i2 / 2;
+        // iinc
+        int i4 = i2;
+        i4++;
+        // imul
+        int i5 = i2 * 2;
+        // ineg
+        int i6 = -i2;
+        // irem
+        int i7 = i2 % 10;
+        // isub
+        int i8 = i - i;
+
+        System.out.println("i:");
+        System.out.println(i);
+
+        System.out.println("i2:");
+        System.out.println(i2);
+
+        System.out.println("i3:");
+        System.out.println(i3);
+
+        System.out.println("i4:");
+        System.out.println(i4);
+
+        System.out.println("i5:");
+        System.out.println(i5);
+
+        System.out.println("i6:");
+        System.out.println(i6);
+
+        System.out.println("i7:");
+        System.out.println(i7);
+
+        System.out.println("i8:");
+        System.out.println(i8);
+    }
+}
