--- src/test/resources/java-test-files/cmjava2023/primitive/long/mathops/Main.java
@@ -1,40 +1,0 @@
-package org.cmjava2023;
-
-public class Main {
-    public static void main(String[] args) {
-        long l = 10;
-        // ladd
-        long l2 = l + l;
-        // ldiv
-        long l3 = l2 / 2;
-        // lmul
-        long l4 = l2 * 2;
-        // lneg
-        long l5 = -l2;
-        // lrem
-        long l6 = l2 % 10;
-        // lsub
-        long l7 = l - l;
-
-        System.out.println("l:");
-        System.out.println(l);
-
-        System.out.println("l2:");
-        System.out.println(l2);
-
-        System.out.println("l3:");
-        System.out.println(l3);
-
-        System.out.println("l4:");
-        System.out.println(l4);
-
-        System.out.println("l5:");
-        System.out.println(l5);
-
-        System.out.println("l6:");
-        System.out.println(l6);
-
-        System.out.println("l7:");
-        System.out.println(l7);
-    }
-}
